+++ library/SolverUtils/Advection/AdvectionFR.cpp
@@ -1215,14 +1215,11 @@ namespace Nektar
                         Vmath::Reverse(nEdgePts, &fluxJumps[0], 1, 
                                        &fluxJumps[0], 1);
                     }
-                    
-                    NekDouble fac = fields[0]->GetExp(n)->EdgeNormalNegated(e) ?
-                        -1.0 : 1.0;
 
                     for (i = 0; i < nEdgePts; ++i)
                     {
-                        if (m_traceNormals[0][trace_offset+i] != fac*normals[0][i] 
-                        || m_traceNormals[1][trace_offset+i] != fac*normals[1][i])
+                        if (m_traceNormals[0][trace_offset+i] != normals[0][i] 
+                        || m_traceNormals[1][trace_offset+i] != normals[1][i])
                         {
                             fluxJumps[i] = -fluxJumps[i];
                         }
@@ -1334,8 +1331,6 @@ namespace Nektar
             int nquad0;
             int nquad1;
             
-            NekDouble fac;
-            
             Array<OneD, NekDouble> auxArray1, auxArray2;
             Array<OneD, LibUtilities::BasisSharedPtr> base;
             
@@ -1417,15 +1412,12 @@ namespace Nektar
                                 fluxN_R[i] = (m_Q2D_e0[n][i]) * fluxN[i];
                             }
                             
-                            fac = fields[0]->GetExp(n)->EdgeNormalNegated(e) ?
-                            -1.0 : 1.0;
-                            
                             for (i = 0; i < nEdgePts; ++i)
                             {
                                 if (m_traceNormals[0][trace_offset+i]
-                                    != fac*normals[0][i] ||
+                                    != normals[0][i] ||
                                     m_traceNormals[1][trace_offset+i]
-                                    != fac*normals[1][i])
+                                    != normals[1][i])
                                 {
                                     fluxN_R[i] = -fluxN_R[i];
                                 }
@@ -1483,15 +1475,12 @@ namespace Nektar
                                 fluxN_R[i] = (m_Q2D_e1[n][i]) * fluxN[i];
                             }
                             
-                            fac = fields[0]->GetExp(n)->EdgeNormalNegated(e) ?
-                            -1.0 : 1.0;
-                            
                             for (i = 0; i < nEdgePts; ++i)
                             {
                                 if (m_traceNormals[0][trace_offset+i]
-                                    != fac*normals[0][i] ||
+                                    != normals[0][i] ||
                                     m_traceNormals[1][trace_offset+i]
-                                    != fac*normals[1][i])
+                                    != normals[1][i])
                                 {
                                     fluxN_R[i] = -fluxN_R[i];
                                 }
@@ -1551,15 +1540,12 @@ namespace Nektar
                                 fluxN_R[i] = (m_Q2D_e2[n][i]) * fluxN[i];
                             }
                             
-                            fac = fields[0]->GetExp(n)->EdgeNormalNegated(e) ?
-                            -1.0 : 1.0;
-                            
                             for (i = 0; i < nEdgePts; ++i)
                             {
                                 if (m_traceNormals[0][trace_offset+i]
-                                    != fac*normals[0][i] ||
+                                    != normals[0][i] ||
                                     m_traceNormals[1][trace_offset+i]
-                                    != fac*normals[1][i])
+                                    != normals[1][i])
                                 {
                                     fluxN_R[i] = -fluxN_R[i];
                                 }
@@ -1620,15 +1606,12 @@ namespace Nektar
                                 fluxN_R[i] = (m_Q2D_e3[n][i]) * fluxN[i];
                             }
                             
-                            fac = fields[0]->GetExp(n)->EdgeNormalNegated(e) ?
-                            -1.0 : 1.0;
-                            
                             for (i = 0; i < nEdgePts; ++i)
                             {
                                 if (m_traceNormals[0][trace_offset+i]
-                                    != fac*normals[0][i] ||
+                                    != normals[0][i] ||
                                     m_traceNormals[1][trace_offset+i]
-                                    != fac*normals[1][i])
+                                    != normals[1][i])
                                 {
                                     fluxN_R[i] = -fluxN_R[i];
                                 }